Added more/better checks
Building again

+++ build.sh
@@ -70,6 +70,8 @@ function build {
     IMAGE_NAME_VERSION=$IMAGE_TAG
     IMAGE_VERSION=${IMAGE_TAG#*:}
 
+    set_picture_paths $IMAGE_TAG
+
     template_go_src main.go main.build.go
 
     [ "$TEMPLATE_CMD" = "CMD" ] && die "build: Missing command in <$TEMPLATE_CMD>"
@@ -89,14 +91,13 @@ function build {
     TIME docker pull $IMAGE_TAG    || true
 
     # Build the compile stage:
-    TIME docker build --target build-env     --cache-from=$STAGE1_IMAGE --tag $STAGE1_IMAGE . ||
-        die "Build failed"
+    TIME docker build --target build-env     --cache-from=$STAGE1_IMAGE --tag $STAGE1_IMAGE . || die "Build failed" 
 
     # Build the runtime stage, using cached compile stage:
     TIME docker build --target runtime-image \
                  --cache-from=$STAGE1_IMAGE \
                  --cache-from=$IMAGE_TAG --tag $IMAGE_TAG . || die "Build failed"
-    echo "CMD=<$TEMPLATE_CMD>"
+    #echo "CMD=<$TEMPLATE_CMD>"
     [ "$TEMPLATE_CMD" = "CMD" ] && die "Missing command in <$TEMPLATE_CMD>"
 
     echo; echo "---- [docker] Checking $IMAGE_TAG command ----------"
@@ -105,32 +106,71 @@ function build {
     ITAG=$(echo $IMAGE_TAG | sed 's?[/:_]?-?g')
     echo; echo "---- [docker] Checking $IMAGE_TAG version ----------"
     docker rm --force name versiontest-$ITAG 2>/dev/null
+    #set -x
     VERSION=$(docker run --rm --name versiontest-$ITAG $IMAGE_TAG $APP_BIN --version 2>&1)
+    #set +x
+    [ -z "$VERSION" ] && die "Failed to create container <versiontest-${ITAG}>"
     echo $VERSION | grep $DATE_VERSION || die "Bad version '$DATE_VERSION' not found in '$VERSION'"
 
     echo; echo "---- [docker] Testing  $IMAGE_TAG ----------"
     let DELAY=LIVE+READY
     [ $DELAY -ne 0 ] && { echo "Waiting for live/ready $LIVE/$READY secs"; sleep $DELAY; }
 
-    CONTAINERNAME=buildtest-$ITAG
-    docker rm --force name $CONTAINERNAME 2>/dev/null
-    # Use default command:
-    #docker run --rm -d --name $CONTAINERNAME -p 8181:$EXPOSE_PORT $IMAGE_TAG $APP_BIN
-    docker run --rm -d --name $CONTAINERNAME -p 8181:$EXPOSE_PORT $IMAGE_TAG
-    CONTAINERID=$(docker ps -ql)
-    curl -sL 127.0.0.1:8181/1 ||
-      die "Failed to interrogate container <$CONTAINERID> $CONTAINERNAME from image <$IMAGE_TAG>"
-    docker stop $CONTAINERID
-    #docker rm $CONTAINERID
+    docker_test_image
+    #kubernetes_test_image
 
     # Push the new versions:
     TIME docker push $STAGE1_IMAGE
     TIME docker push $IMAGE_TAG
-
-    test_kubernetes_images 
 }
 
-function test_kubernetes_images {
+function docker_test_image {
+    CONTAINERNAME=buildtest-$ITAG
+    docker rm --force name $CONTAINERNAME 2>/dev/null
+
+    # Use default command:
+    #docker run --rm -d --name $CONTAINERNAME -p 8181:$EXPOSE_PORT $IMAGE_TAG $APP_BIN
+    docker run --rm -d --name $CONTAINERNAME -p 8181:$EXPOSE_PORT $IMAGE_TAG
+    CONTAINERID=$(docker ps -ql)
+
+    curl -sL 127.0.0.1:8181/1 ||
+      die "Failed to interrogate container <$CONTAINERID> $CONTAINERNAME from image <$IMAGE_TAG>"
+
+    TXT_PATH="${PICTURE_PATH_BASE}.txt"
+    [ ! -f $TXT_PATH ] && die "No such txt file <$TXT_PATH>"
+    PNG_PATH="${PICTURE_PATH_BASE}.png"
+    [ ! -f $PNG_PATH ] && die "No such png file <$PNG_PATH>"
+
+    #set -x
+    #CMD="curl -sL 127.0.0.1:8181/${TXT_PATH} | wc -c"
+    CMD="curl -sL 127.0.0.1:8181/${TXT_PATH}"
+    CURL_TXT_SIZE=$($CMD | wc -c)
+    [ -z "$CURL_TXT_SIZE" ] && die "curl command failed <$CMD>"
+    #set +x
+
+    #CMD="wc -c < ${TXT_PATH}"
+    CMD="cat ${TXT_PATH}"
+    FILE_TXT_SIZE=$($CMD | wc -c)
+    [ -z "$FILE_TXT_SIZE" ] && die "wc command failed <$CMD>"
+
+    [ "$CURL_TXT_SIZE" != "$FILE_TXT_SIZE" ] && die "Different text image sizes [ $CURL_TXT_SIZE != $FILE_TXT_SIZE ] ($TXT_PATH)"
+
+    CMD="curl -sL 127.0.0.1:8181/${PNG_PATH}"
+    CURL_PNG_SIZE=$($CMD | wc -c)
+    [ -z "$CURL_PNG_SIZE" ] && die "curl command failed <$CMD>"
+
+    CMD="wc -c < ${PNG_PATH}"
+    CMD="cat ${PNG_PATH}"
+    FILE_PNG_SIZE=$($CMD | wc -c)
+    [ -z "$FILE_PNG_SIZE" ] && die "wc command failed <$CMD>"
+
+    [ "$CURL_PNG_SIZE" != "$FILE_PNG_SIZE" ] && die "Different PNG image sizes [ $CURL_PNG_SIZE != $FILE_PNG_SIZE ] ($PNG_PATH)"
+
+    docker stop $CONTAINERID
+    #docker rm $CONTAINERID
+}
+
+function kubernetes_test_image {
     # NO USE as this CAN ONLY BE DONE AFTER push
 
     JOBNAME=kubejobtest-$ITAG
@@ -178,10 +218,19 @@ function test_kubernetes_images {
     sleep 2
 
     curl -sL 127.0.0.1:8181/1 || {
-        kill -9 $PID
-        kubectl delete pod/$PODNAME
+        #kubectl delete pod/$PODNAME
+	echo "----- ERROR"
+        echo "Test then 'kubectl delete pod/$PODNAME'"
+        echo "Test then 'kill -9 $PID' # port-forward"
         die "Failed to interrogate pod <$PODNAME> from image <$IMAGE_TAG>"
     }
+
+    CURL_TXT_SIZE=$(curl -sL 127.0.0.1:8181/${PICTURE_PATH_BASE}.txt | wc -c)
+    FILE_TXT_SIZE=$(wc -c < ${PICTURE_PATH_BASE}.txt)
+    [ $CURL_TXT_SIZE -ne $FILE_TXT_SIZE ] && die "Different text image sizes [ $CURL_TXT_SIZE -ne $FILE_TXT_SIZE ]"
+    CURL_PNG_SIZE=$(curl -sL 127.0.0.1:8181/${PICTURE_PATH_BASE}.Pgt | wc -c)
+    FILE_PNG_SIZE=$(wc -c < ${PICTURE_PATH_BASE}.txt)
+    [ $CURL_PNG_SIZE -ne $FILE_PNG_SIZE ] && die "Different PNG image sizes [ $CURL_PNG_SIZE -ne $FILE_PNG_SIZE ]"
 
     # NEED TO KILL POD
     kill -9 $PID
@@ -189,13 +238,8 @@ function test_kubernetes_images {
 
 }
 
-function template_dockerfile {
-    echo "FN: template_dockerfile $*"
-    # e.g. template_dockerfile mjbright/ckad-demo:1 scratch 80 ["/app/demo-binary","--listen",":80","-l","0","-r","0"]
+function set_picture_paths {
     IMAGE_TAG=$1; shift
-    FROM_IMAGE=$1; shift
-    EXPOSE_PORT=$1; shift
-    TEMPLATE_CMD="$*"; set --
 
     PICTURE_TYPE=""
     case $IMAGE_TAG in
@@ -207,12 +251,12 @@ function template_dockerfile {
 
     COLOUR=""
     case $IMAGE_TAG in
-        *:1) COLOUR="blue";;
-        *:2) COLOUR="red";;
-        *:3) COLOUR="green";;
-        *:4) COLOUR="cyan";;
-        *:5) COLOUR="yellow";;
-        *:6) COLOUR="white";;
+        *:1|*:alpine1) COLOUR="blue";;
+        *:2|*:alpine2) COLOUR="red";;
+        *:3|*:alpine3) COLOUR="green";;
+        *:4|*:alpine4) COLOUR="cyan";;
+        *:5|*:alpine5) COLOUR="yellow";;
+        *:6|*:alpine6) COLOUR="white";;
         *)   die "Unknown image tag: <$IMAGE_TAG>";;
     esac
 
@@ -221,6 +265,22 @@ function template_dockerfile {
 
     [ ! -f "${PICTURE_PATH_BASE}.png" ] && die "No such file <${PICTURE_PATH_BASE}.png>"
     [ ! -f "${PICTURE_PATH_BASE}.txt" ] && die "No such file <${PICTURE_PATH_BASE}.txt>"
+}
+
+function check_vars_set {
+    for var in $*; do
+        eval val=\$var
+        [ -z "$val" ] && die "Variable \$var is unset"
+    done
+}
+
+function template_dockerfile {
+    #echo "FN: template_dockerfile $*"
+    # e.g. template_dockerfile mjbright/ckad-demo:1 scratch 80 ["/app/demo-binary","--listen",":80","-l","0","-r","0"]
+    IMAGE_TAG=$1; shift
+    FROM_IMAGE=$1; shift
+    EXPOSE_PORT=$1; shift
+    TEMPLATE_CMD="$*"; set --
 
     #echo "EXPOSE_PORT=$EXPOSE_PORT"
     [ "$TEMPLATE_CMD" = "CMD" ] && die "template_dockerfile: Missing command in <$TEMPLATE_CMD>"
@@ -234,15 +294,18 @@ function template_dockerfile {
         *)  die "Unknown FROM_IMAGE type <$FROM_IMAGE>";;
     esac
 
+    check_vars_set FROM_IMAGE EXPOSE_PORT STAGE1_BUILD TEMPLATE_CMD
+
     sed  < templates/Dockerfile.tmpl > Dockerfile \
         -e "s/__FROM_IMAGE__/$FROM_IMAGE/" \
         -e "s/__EXPOSE_PORT__/$EXPOSE_PORT/" \
         -e "s/__STAGE1_BUILD__/$STAGE1_BUILD/" \
         -e "s?__TEMPLATE_CMD__?$TEMPLATE_CMD?" \
-        -e "s?__PICTURE_BASE__?$PICTURE_BASE?" \
-        -e "s?__PICTURE_PATH_BASE__?$PICTURE_PATH_BASE?" \
 
-    grep __ Dockerfile && die "Uninstantiated variables in Dockerfile"
+        #-e "s?__PICTURE_BASE__?$PICTURE_BASE?" \
+        #-e "s?__PICTURE_PATH_BASE__?$PICTURE_PATH_BASE?" \
+
+    grep -v "^#" Dockerfile | grep __ && die "Uninstantiated variables in '${BUILD_SRC}'"
     mkdir -p tmp
 
     DFID=$(echo $IMAGE_TAG | sed -e 's/\//_/g')
@@ -268,14 +331,18 @@ function push {
 }
 
 function build_and_push {
-    #IMAGE_TAG=$1; shift
+    # build_and_push $IMAGE scratch $PORT $CMD
+    IMAGE_TAG=$1; shift
+    FROM_IMAGE=$1; shift
+    PORT=$1; shift
+    CMD=$1; shift
     #build $IMAGE_TAG
     #push $IMAGE_TAG
 
     #[ "$TEMPLATE_CMD" = "CMD" ] && die "build: Missing command in <$TEMPLATE_CMD>"
     echo $4
-    build $*
-    push  $*
+    build $IMAGE_TAG $FROM_IMAGE $PORT $CMD
+    push  $IMAGE_TAG $FROM_IMAGE $PORT $CMD
 }
 
 # START: TIMER FUNCTIONS ================================================
@@ -293,13 +360,23 @@ function TIMER_STOP {
 function TIME {
     CMD=$*
 
+    #FILE_SUFFIX=$(echo $CMD | sed 's/\(\"| |\,|\>|\<|\/)*/_/g' | tr "'" "_") 
+    FILE_SUFFIX=$(echo $CMD | tr -s  "\"\\\/ '<>,:" "_")
+    CMD_OP=tmp/cmd.op.$FILE_SUFFIX
+    #echo CMD_OP=$CMD_OP
+    touch $CMD_OP || die "Failed to touch <$CMD_OP>"
+
     CMD_TIME=$(date +%Y-%b-%d_%02Hh%02Mm%02S)
-    echo; echo "---- [$CMD_TIME] $CMD"
+    echo "---- [$CMD_TIME] $CMD"
     TIMER_START
-    $CMD; RET=$?
+    $CMD > $CMD_OP 2>&1; RET=$?
     TIMER_STOP
     echo "Took $TOOK secs [${HRS}h${MINS}m${SECS}]"
-    [ $RET -ne 0 ] && echo "ERROR: returned $RET"
+    [ $RET -ne 0 ] && {
+        pwd
+        cat $CMD_OP
+        die "ERROR: returned $RET"
+    }
     return $RET
 }
 
@@ -320,16 +397,18 @@ function template_go_src {
     SRC=$1;       shift
     BUILD_SRC=$1; shift
 
+    check_vars_set DATE_VERSION IMAGE_NAME_VERSION IMAGE_VERSION PICTURE_PATH_BASE
+
     sed < ${SRC} > ${BUILD_SRC}  \
            -e "s/__DATE_VERSION__/$DATE_VERSION/" \
            -e "s?__IMAGE_NAME_VERSION__?$IMAGE_NAME_VERSION?" \
            -e "s/__IMAGE_VERSION__/$IMAGE_VERSION/" \
            -e "s?__PICTURE_PATH_BASE__?$PICTURE_PATH_BASE?" \
 
-    ls -altr ${SRC} ${BUILD_SRC}
+    #ls -altr ${SRC} ${BUILD_SRC}
     [ ! -s ${BUILD_SRC} ] && die "Empty ${BUILD_SRC} !!"
 
-    grep __ ${BUILD_SRC} && die "Uninstantiated variables in '${BUILD_SRC}'"
+    grep -v "^#" ${BUILD_SRC} | grep __ && die "Uninstantiated variables in '${BUILD_SRC}'"
 }
 
 function build_and_push_tags {
@@ -358,9 +437,10 @@ function build_and_push_tags {
 IMAGE_NAME_VERSION=""
 IMAGE_VERSION=""
 
-template_go_src main.go main.build.go
+#set_picture_paths $IMAGE_TAG
+#template_go_src main.go main.build.go
+#TIME check_build main.build.go
 
-TIME check_build main.build.go
 #die "OK"
 
 ## -- Args: -------------------------------------------------------------
@@ -385,13 +465,17 @@ done
 
 [ -z "$TAGS" ] && TAGS="1"
 
-echo "Building repos<$REPO_NAMES> tags<$TAGS>"
+echo "Building repos<$REPO_NAMES> tags<"$TAGS">"
 
 ## -- Main: -------------------------------------------------------------
 
 # Incremental builds:
 
-docker login
+docker login > ~/tmp/docker.login.op 2>&1 || {
+    cat ~/tmp/docker.login.op
+    die "Failed to login to Docker Hub"
+}
+#kubectl get nodes || die "Failed to access cluster"
 
 TIMER_START; START0_S=$START_S
